requirements.txt
streamlit==1.38.0
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
Pillow==10.4.0
moviepy==1.0.3
imageio-ffmpeg==0.5.1
soundfile==0.12.1
librosa==0.10.2.post1
numba==0.59.1
llvmlite==0.42.0
nltk==3.9.1
